Test dependency/ordering fixes
Remove priorities from tests.  The hints are mostly unnecessary due to
the high default parallelism (10) in the test runner.  The cost of
potential confusion about the ordering is also significant.

Remove incorrect comments about the cause of the incorrect ordering.
The cause has been identified as issue 1379 in TestNG which results from
mixing parallel with dependsOnMethods.  Disable parallel execution until
this bug is fixed.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>