Run updateProject after updateProjectSettingsYaml
The reason the build settings were being reset was not due to
initialization in cancelBuild, but due to being overwritten by
updateProjectSettingsYaml restoring settings from before updateProject
was called (because getProjectSettingsYaml ran before updateProject).
I'm not sure why I thought it was cancelBuild-related.

Improve the test ordering to avoid this.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>