Update badges
Remove Travis CI badge since I don't plan to use it for this project.
Given that the test suite makes the AppVeyor API do a lot of work, I'd
rather not run a second CI.

Remove the coverage badge since I don't plan to use it.  swagger-codegen
produces a lot of code and I don't plan to get reasonable coverage on
it, since it is mostly out of my control.

Switch to AppVeyor badge for AppVeyor status.  Since it is an
AppVeyor-centric project, having the distinctive badge is worth it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>